Committing Polished Python A/B Test example to master branch.
Project is nearly finished.
